fix: fixed blank

+++ backend/blank/schemas.py
@@ -2,7 +2,7 @@ from pydantic import BaseModel
 
 
 class CreaetBlankDTO(BaseModel):
-    user_1: int
-    user_2: int
-    user_3: int
-    user_4: int
+    user_1: float
+    user_2: float
+    user_3: float
+    user_4: float
